+++ CHANGELOG.md
@@ -20,6 +20,15 @@ The **patch** part changes is incremented if multiple releases happen the same m
 
 ### Security
 
+## [22.06.1] 2022-06-06
+
+### Changed
+
+* evocheck: upstream release 22.06.1
+* minifirewall: upstream release 22.06
+* mysql: evomariabackup release 22.06.1
+* mysql: reorganize evomariabackup to use mtree instead of our own dir-check
+
 ## [22.06] 2022-06-03
 
 ### Added